fix(capture): two P1s my own fixes introduced, and the triage lane's label
The re-look-after-the-lock I added an hour ago trusted the ledger's
organization id. ArchiveOrganization archives the row, its domains AND its
relationships — which is precisely why the dedupe found nothing and this path
ran at all — so attaching on that stale id would recreate employment on a
company somebody deliberately retired. It now checks the organization is live,
under a row lock so the answer cannot change beneath the insert.

The rollback conversion I added in the same round could fail on a race. A
triage worker caught between binding its organization and finishing still holds
a live row, and turning that one into an `organization` read can violate
uq_site_read_org_inflight when the same organization already has an active read
for that seed. Only FINISHED bound reads convert now, and only where no live
organization read would collide; an unfinished one is discarded with the rest,
having evidenced nothing.

The debug report labelled triage calls `other`, so the lane it was added to
expose could not be told apart from the rest.

ensure.go crossed the 500-line cap on the way, so attaching a person to a
company — whether the company is still one to attach to, and the edge itself —
moves to employmentedge.go, and the settled-verdict branch becomes a named
function rather than three levels of nesting.
